Struct lib_wd::wdinfo::WDInfo

pub struct WDInfo {
    pub pg: PostgreSQL,
    // some fields omitted
}

WDInfo struct

pg as PostgreSQL connection (public)

temporary schema, default wdinfo

temporary table for PostgreSQL::import_data, default _file_st

table structure for temporary table (string)

inserted timestamp (not used?)

key variable for tables, default fullpath

prefix string for 192.168.1.241, default //192.168.1.241/public/

prefix string for 192.168.1.243, default //192.168.1.243/

Fields
pg: PostgreSQL
Implementations

impl WDInfo

pub fn wdrefresh(
&mut self,
path: &str,
skm: &str,
tbl: &str
) -> Result<(), &'static str>

WD refresh: refresh PostgreSQL table based on files on WD net drive

* Like one interface of this library
* Calls self.fs_import_pg and self.wdinfo_refresh

pub fn wdsync(
&mut self,
skm: &str,
newtbl: &str,
oldtbl: &str
) -> Result<(), &'static str>

WD sync: sync WD net drive and PostgreSQL table based on records in tables

* Like one interface of this library
* Calls self.wdinfo_compare and self.wdinfo_sync

pub fn new(
host: String,
username: String,
password: String,
database: String
) -> Result<Self, &'static str>

WDInfo initialization, other than default

* use customer defined PostgreSQL connection arguments
* Other private elements are still using default

pub fn new_special(
host: String,
username: String,
password: String,
database: String,
tmp_skm: String,
tmp_tbl: String,
tbl_str: String,
insdt: String,
keyvar: String,
pre241: String,
pre243: String
) -> Result<Self, &'static str>

WDInfo initialization, other than default

* use customer defined PostgreSQL connection arguments
* And other private elements

pub fn last_insert_dt(
&mut self,
skm: &str,
tbl: &str
) -> Result<String, &'static str>

Check last time refresh of PostgreSQL table based on WD net drive

* (not used?)

pub fn fs_import_pg(&mut self, path: &str) -> Result<(), &'static str>

Check WD net drive, import FileStatus into PostgreSQL temporary table

* The time consuming function: log start and finish status
* call FileStatus::get_file_status_under_folder with dilimeter "|", then
* call self.pg.import_data with query using dilimeter "|"

pub fn wdinfo_update(
&mut self,
skm: &str,
tbl: &str
) -> Result<(), &'static str>

update table using temporary table

* Not used?

pub fn wdinfo_refresh(
&mut self,
skm: &str,
tbl: &str
) -> Result<(), &'static str>

refresh table using temporary table

* truncate table
* insert into table using now()::timestamp

pub fn wdinfo_compare(
&mut self,
skm: &str,
newtbl: &str,
oldtbl: &str
) -> Result<Vec<String>, &'static str>

compare tables for records represent different WD net drive for sync later

* call self.pg.query with return Vec<Row>
* convert return to Vec<String>

pub fn wdinfo_sync_one(
&mut self,
skm: &str,
newtbl: &str,
oldtbl: &str,
fullpath: &str
) -> Result<(), &'static str>

sync WD net drive for one file

* call FileStatus::copy_file
* update table for this one record (append)
* Consider file name with single quotation (PostgreSQL use two single quotation represent one single quotation)
* path in tables start with //IP_address; convert to /mnt for copying file

pub fn wdinfo_sync(
&mut self,
skm: &str,
newtbl: &str,
oldtbl: &str,
pathlist: Vec<String>
) -> Result<(), &'static str>

sync WD net drive for giving file list

* call self.wdinfo_sync_one
* time consuming process, with log of start and finished